__pycache__/requirements.txt
streamlit==1.35.0
google-generativeai==0.3.0
supabase==2.5.0
pdfplumber==0.11.0
pandas==2.2.3
openpyxl==3.1.5
plotly==5.24.1
psycopg2-binary==2.10.1
